requirements.txt
absl-py==2.1.0
asttokens==2.0.8
astunparse==1.6.3
backcall==0.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.0
Flask==3.0.3
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.11.0
idna==3.10
ipykernel==6.16.0
ipython==8.5.0
itsdangerous==2.2.0
jedi==0.18.1
Jinja2==3.1.4
jupyter-core==4.11.1
jupyter_client==7.3.5
keras==3.4.1
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
nest-asyncio==1.5.5
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
pillow==10.4.0
prompt-toolkit==3.0.31
protobuf==4.25.5
psutil==5.9.2
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pywin32==304
pyzmq==24.0.1
requests==2.32.3
rich==13.8.1
six==1.16.0
stack-data==0.5.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tornado==6.2
traitlets==5.4.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.5
Werkzeug==3.0.4
wrapt==1.16.0
